Fixed a int missmatch

+++ 3B2/3b2_stddev.c
@@ -307,8 +307,8 @@ static void tod_resync(UNIT *uptr)
         if (delta > MIN_DIFF && delta < MAX_DIFF) {
             catchup_ticks = (uint32) delta * CLK_TPS;
             sim_debug(EXECUTE_MSG, &tod_dev,
-                      "Catching up with a delta of %ld seconds (%d ticks).\n",
-                      delta, catchup_ticks);
+                      "Catching up with a delta of %lld seconds (%d ticks).\n",
+                      (long long) delta, catchup_ticks);
             while (catchup_ticks-- > 0) {
                 tod_tick(&tod_unit);
             }
